Reverting eltype(::LoopValue) change from last commit.

+++ src/add_loads.jl
@@ -77,5 +77,5 @@ struct LoopValue end
 @inline SIMDPirates.vload(::LoopValue, i::Tuple{_MM{W}}, ::Unsigned) where {W} = SVec(SIMDPirates.vrangeincr(Val{W}(), @inbounds(i[1].i), Val{1}()))
 @inline VectorizationBase.load(::LoopValue, i::Integer) = i + one(i)
 @inline VectorizationBase.load(::LoopValue, i::Tuple{I}) where {I<:Integer} = @inbounds(i[1]) + one(I)
-@inline Base.eltype(::LoopValue) = Int
+@inline Base.eltype(::LoopValue) = Int8
